% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalomicsdbR.R
\docType{data}
\name{clinicalomicsdbR}
\alias{clinicalomicsdbR}
\title{clinicalomicsdbR object}
\format{
An object of class \code{list} of length 2.
}
\usage{
clinicalomicsdbR
}
\description{
clinicalomicsdbR object
}
\keyword{datasets}
